PR#14287 comments applied
